requirements.txt
Flask==3.0.0
pypdf==3.17.4
python-docx==1.1.0
langdetect==1.0.9
gunicorn==21.2.0
werkzeug==3.0.1